Gemini/Langchain/requirements.txt
langchain
python-dotenv
streamlit
langchain_community
langchain_groq
fastapi
langserve
sse_starlette
langchain-chroma
langchain_huggingface
bs4
